# packages/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: HashHive Control API
  description: |
    Machine-readable REST API for CLI tooling, automation platforms, and CI
    integrations. Authenticated by per-user API keys (format `cst_*`),
    paginated with `offset`/`limit`, and emits RFC 9457 problem-details on
    errors.
  version: 1.0.0
servers:
  - url: /api/v1/control
    description: Control API base path
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: List projects the user belongs to
      tags: [Projects]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProjects"
        "401":
          $ref: "#/components/responses/AuthError"
  /projects/{id}:
    get:
      operationId: getProject
      summary: Get a single project
      tags: [Projects]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Project record
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/me:
    get:
      operationId: getMe
      summary: Authenticated user's profile
      tags: [Users]
      security:
        - BearerApiKey: []
      responses:
        "200":
          description: User record
        "401":
          $ref: "#/components/responses/AuthError"
  /users:
    get:
      operationId: listUsers
      summary: List users (admin on active project)
      tags: [Users]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated users
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{id}:
    get:
      operationId: getUser
      summary: Get a single user
      tags: [Users]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: User record
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /resources/{kind}:
    get:
      operationId: listResources
      summary: List resources of a given kind
      tags: [Resources]
      security:
        - BearerApiKey: []
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [wordlists, rulelists, masklists]
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated resources
  /resources/{kind}/{id}:
    get:
      operationId: getResource
      summary: Get a single resource
      tags: [Resources]
      security:
        - BearerApiKey: []
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [wordlists, rulelists, masklists]
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Resource record
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tasks:
    get:
      operationId: listTasks
      summary: List tasks for a given campaign
      tags: [Tasks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - in: query
          name: campaignId
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated tasks
  /tasks/{id}:
    get:
      operationId: getTask
      summary: Get a task
      tags: [Tasks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Task record
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: cst_<userId>_<random>
  parameters:
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
    ProjectIdHeader:
      in: header
      name: X-Project-Id
      description: Active project the request operates against
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details envelope
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path:
                type: string
              code:
                type: string
              message:
                type: string
    PaginatedProjects:
      type: object
      required: [items, total, offset, limit]
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
  responses:
    AuthError:
      description: Authentication failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Caller lacks permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Target resource was not found or is not visible to the caller
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...